Refactor: 게임 N round 진행 함수 CarRaceGameUtility에 위임
게임 N round 진행 함수 CarRaceGameUtility에 위임하여 컨트롤러가 컨트롤 역할만 하도록 하였다.
